Move launch view to view folder
