+++ desktop/packages/mullvad-vpn/src/renderer/components/views/index.ts
@@ -1,4 +1,5 @@
 export * from './app-info';
 export * from './app-upgrade';
+export * from './launch';
 export * from './changelog';
 export * from './settings';
