+++ desktop/packages/mullvad-vpn/src/renderer/components/views/launch/index.ts
@@ -1,0 +1,1 @@
+export * from './Launch';
